Fix a test written against OpenJDK to work with Oracle's JDK.
Using java 1.7.0_25, the ZipFileSystem appears to have a dependency on the real
file system, even when the zip file exists with JimFS. By creating the test zip
at /out.zip, ZFS would try to write a temp file under /, which would trigger a
permission denied error. This changes the test to use java.io.tmpdir, even when
working with JimFS, so the JDK's ZipFileSystem does the right thing.