For issue #26, do not document types with private visibility.
While private constructors may be used to restrict where a class may be
instantiated, the more common case is to restrict usage to the defining file
(this ambiguity could also be addressed by simply not exporting the type from
a module).

This commit is a bit of a temporary change. There should be a config option to
control what level of visibility to document (public, protected, or private).
Furthermore, logic needs to be added for the inherited case, where a visibility
has been set at the file level.