Search for property/function documentation on implemented/extended types.
Documentation for block comments, parameters, and return types will be used from
the nearest assignable type:
- check jsdoc present directly on the nominal type
- check the jsdoc of interfaces directly implemented by the type
-- a recursive search of the super-interfaces of directly implemented interfaces
- check the jsdoc of the super-class (if the nominal type is a class)
-- a recursive search of the interfaces directly implemented by the super-class

Individual parameters may not be overridden; it's an all or nothing affair
(patches are welcome to improve this).

Throws and deprecation declarations are not inherited. Each type must directly
declare these annotations for it to show up in generated documentation. This is
the case even or "empty" comments of the form "/** @override */".

Fixes #30.