Change how the main page scripts are loaded
Dynamically generate the script tags via soy instead of embedded js
so they can use the defer tag.

Move loading of the main index to <head> to ensure it is loaded before
the main app logic. There's more cleanup that can be done here to
handle the case where, if loaded via <body> and types.js gets executed
last, it will call into the app to initialize the menu and search box
after the fact. But for now, using script with defer works.