Upgrade the closure compiler dependency to v20170521
This is a pretty big version jump and caused a lot of internal changes due to
changes in compiler behavior. Notably, the compiler does more inlining before
custom passes are run, so dossier has to do more work to track aliases (refer to
diff for details).

Also, the compiler seems to be making more assumptions about ES6 structure and
what types of behavior is supported (e.g. default exports don't seem to be
supported by the compiler anymore, so Dossier won't be able to document those
properly)