Get things building with Bazel 0.9.0
This turned into a yak shaving exercise. Closure build tools 0.4.1 don't work
with Bazel 0.9.0, so I had to upgrade the build tools to 0.4.2. This upgraded
version includes a stricter linter, so I had to fix a bunch of type warnings in
the JS code.

0.4.2 was also crashing a lot from opening too many files at once. This was
caused by passing the entire Closure library in the compile step (800+ files).
To fix this, I added some custom library rules that only include the .js files
from Closure that are actually used here (~140). This was tedious to setup, but
builds are faster and more stable now.